move some tests from weight-proof to proof-of-space
